9.11 1：为各个目录增加init文件 2：完成上传测试用例列表功能(未对内容校验)

+++ common/most_common_method.py
@@ -1,7 +1,5 @@
-import datetime
-import os
-import time
 from config import TESTCASE_XLSX_PATH
+from . import datetime, xlrd, time, os
 
 
 class NullObject:
@@ -19,8 +17,21 @@ def clear_download_xlsx(method='download'):
     path_list = os.listdir(dir_path)
     target_time = int(time.time())-300
     for path in path_list:
-        __path = dir_path + '/' + path
-        path_time = int(os.path.getctime(__path))
-        print('path_time:', path_time, target_time, __path)
-        if path_time <= target_time:
-            os.remove(__path)
+        if '.py' not in path:
+            __path = dir_path + '/' + path
+            path_time = int(os.path.getctime(__path))
+            print('path_time:', path_time, target_time, __path)
+            if path_time <= target_time:
+                os.remove(__path)
+        else:
+            continue
+
+
+def read_xlsx(table):
+    book = xlrd.open_workbook(table)
+    row_list = []
+    for s in range((len(book.sheets()))):
+        sheet = book.sheets()[s]
+        for row in range(sheet.nrows):
+            row_list.append(sheet.row_values(row))
+    return row_list
